Started work on the new asset menu!
Now there's these cool things called folders! It's a totally new
invention, I think it'll do really well with those tiktok kids.

The way in which assets are displayed can also be adjusted now, with
both the button type and scale being variable.

This is primarily to fix #268, as the 3D previews were causing
random crashes in v0.1.x when going through the pages.

This also fixes #299, as tooltips now only show descriptions. I plan
to add a right-click option to show the full details of an asset.

This commit also contributes to #252, since tokens are now shown in
the UI as "Tiles / Tokens". The ability for tokens to be placed in
player's hands has not been implemented yet.