Pieces now use distinct 'modes' to define their behaviour.
The general idea is that it is a state machine, where pieces can
transfer between modes depending on the mode that it is currently in.

The current modes are limbo, sleep, normal, and locked. There will
be modes added in the future for flying (when a player is picking up
a piece), and hovering (when a piece is in a player's hand).

The hope is that this will better define what RPCs are sent in each
mode. For example, normal state updates will be sent in normal mode,
but not when flying or hovering.

This has not been reflected in room states as of yet, which will be
done in the near future.