Added RPCs for switching between the new piece modes.
Also, I have FINALLY figured out why box shapes have weird collision,
the cause of #50. I found this out since when unlocking the radio in
the test room, it does not actually settle, thus it does not change
to the sleep mode.

It's due to the fact that we are scaling the collision shape nodes,
not changing the shape resources directly... :|

Wish I would have figured this out, like, two years ago... >:(
So yeah, gonna have to change how scaling works entirely.