Collision shapes are no longer scaled or rotated.
This is so the game conforms to the doc's recommendation to avoid
scaling or rotating the collision shape node itself, so that the
physics engine can use specific optimisations, and to avoid weird
physics issues.

This does mean however, that the shape itself needs to be scaled
(and potentially rotated). This may hurt performance for large
meshes, or when scaling large numbers of pieces, so fingers crossed
the performance hit isn't too big - if it is, then we can always
revert this commit, or scale the shape in C++ instead of GDScript.

Now, this may come as a plot twist, but I'm not 100% sure if this
actually is the fix for #50. The radio still does it's little jiggle
when it lands on the table...

BUT! Because this took literally all day, and frankly I feel like
I need some sort of reward for going through the complete ball-ache
that is physics debugging, I'm going to say that this commit does
indeed fix #50, as I've tried landing the dice a bunch of times at
weird angles, and they DON'T do a jiggle as they are coming to rest.
So it was either this commit, or something else along the line, that
must have fixed that! :D

However, that still leaves the radio issue... well, just you wait
until this next spicy commit... (if the fix I've found actually
works, that is...)