Fixed radio jitter issue by adjusting the COM of tables.
... this took me until three in the flibbin morning to figure out.

It got to the point to where I started a new project, and slowly
started to re-create the same test senario in it. I tried literally
everything, from changing the physics FPS, to different masses, to
different box shape sizes. But I just happened to change the position
of the "floor", and BAM. No more jitter!

So I've put the centre-of-mass of the tables at exactly -10 below
y = 0, closer to the surface. This has eliminated the jitter for
both the poker table and the picnic table.

However, this does mean that the table_transform property in room
states now needs to be ignored if the table has not been flipped yet,
since Tabletop Club v0.1.x sets this value regardless of whether the
table is flipped or not. It should not make a difference if the
table is stationary, but it will change the position of the table if
the state captured a table mid-flight. But I'll take a slightly wonky
position for a table that's been flipped over micro-jitters any day
of the week. Including Wednesdays. Which is today. I started working
on this on Tuesday. :|

As mentioned in a TODO comment, this might still potentially be an
issue for imported tables, since they still use the same COM adjustment
system as objects - so chances are, their COM will be much lower than
the surface. It might be worth having a consistent COM for all tables
like (0, -10, 0) or (0, 0, 0) so that ALL tables avoid the jitter
issue that has just been solved by the default tables.