Switched back to single-safe thread model for rendering.
After a ton of debugging and testing, the reason the physics
interpolation warnings keep appearing is because the rendering model
was set to multi-threaded, and after testing it multiple times,
switching to the single-safe model seems to have stopped the warnings,
while also not causing the game to crash when importing custom assets,
thankfully!

See this issue for more information:
godotengine/godot#97175

NOTE: This setting seems to cause the window that Gut spawns to
be permanently black, although, if the process is stopped manually,
the test results can still be seen in the editor.